+++ example/main/example_restconf.c
@@ -61,6 +61,7 @@ static const char Pad64 = '=';
  *
  * Start restconf with -- -m <yang> -M <namespace>
  * Mount this yang on mountpoint
+ * Note module-set hard-coded to "mylabel"
  */
 static char *_mount_yang = NULL;
 static char *_mount_namespace = NULL;