+++ lib/clixon/clixon_yang_module.h
@@ -79,6 +79,6 @@ yang_stmt *yang_find_module_by_name_revision(yang_stmt *yspec, const char *name,
 yang_stmt *yang_find_module_by_name(yang_stmt *yspec, char *name);
 int        yang_metadata_annotation_check(cxobj *x, yang_stmt *ymod, int *ismeta);
 int        yang_metadata_init(clixon_handle h);
-int        yang_lib2yspec(clixon_handle h, cxobj *yanglib, char *mntpnt, yang_stmt *yspec);
+int        yang_lib2yspec(clixon_handle h, cxobj *yanglib, char *mntpnt, char *domain, yang_stmt *yspec);
 
 #endif  /* _CLIXON_YANG_MODULE_H_ */